NF: Move TagsDialog to its own package
